Update to questions
I've updated the questions, removing some and adding 2. I attempted to
change the id numbers as well, but only for questions that were added.
For questions I removed, I didn't update the ids, as I figured it would
be easiest to leave them as is, so that we don't need to change anything
on the server side.